requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-google-genai
google-generativeai
torch 
transformers